+++ CHANGES.rst
@@ -7,6 +7,7 @@ What's new
 Bug fixes
 ~~~~~~~~~
 * Raise a meaningful error messages when the output grid has no chunks with `parallel=True` (:issue:`299`, :pull:`304`). By `Pascal Bourgault <https://github.com/aulemahal>`_.
+* Correct guess of output chunks for the :``SpatialAverager``.
 
 0.8.1 (2023-09-05)
 ------------------
@@ -14,7 +15,6 @@ Bug fixes
 Bug fixes
 ~~~~~~~~~
 * Change import to support shapely 1 and 2.
-
 
 0.8.0 (2023-09-01)
 ------------------